refactor: change naming
