--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/mariatrejer/RS School/React/React-2024-Q3/src/App.tsx":"1","/Users/mariatrejer/RS School/React/React-2024-Q3/src/main.tsx":"2","/Users/mariatrejer/RS School/React/React-2024-Q3/src/vite-env.d.ts":"3","/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/main/Card.tsx":"4","/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/main/CardsWrapper.tsx":"5","/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/searchForm/SearchForm.tsx":"6","/Users/mariatrejer/RS School/React/React-2024-Q3/src/model/TypesStarWars.ts":"7","/Users/mariatrejer/RS School/React/React-2024-Q3/src/pages/MainPage.tsx":"8","/Users/mariatrejer/RS School/React/React-2024-Q3/src/services/ApiService.ts":"9","/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/header/Header.tsx":"10","/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/errorBoundary/ErrorBoundary.test.tsx":"11"},{"size":379,"mtime":1751973040809,"results":"12","hashOfConfig":"13"},{"size":242,"mtime":1751973040816,"results":"14","hashOfConfig":"13"},{"size":38,"mtime":1723583099409,"results":"15","hashOfConfig":"13"},{"size":676,"mtime":1752005250303,"results":"16","hashOfConfig":"13"},{"size":529,"mtime":1752005303129,"results":"17","hashOfConfig":"13"},{"size":1612,"mtime":1752140343281,"results":"18","hashOfConfig":"13"},{"size":263,"mtime":1752005250304,"results":"19","hashOfConfig":"13"},{"size":1873,"mtime":1752005340319,"results":"20","hashOfConfig":"13"},{"size":821,"mtime":1751993648876,"results":"21","hashOfConfig":"13"},{"size":761,"mtime":1752062419212,"results":"22","hashOfConfig":"13"},{"size":1010,"mtime":1752685578035,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","suppressedMessages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1rzmn5y",{"filePath":"27","messages":"28","suppressedMessages":"29","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","suppressedMessages":"32","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","suppressedMessages":"35","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","suppressedMessages":"38","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","suppressedMessages":"41","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","suppressedMessages":"44","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"45","messages":"46","suppressedMessages":"47","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","suppressedMessages":"50","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"51","messages":"52","suppressedMessages":"53","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"54","messages":"55","suppressedMessages":"56","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/mariatrejer/RS School/React/React-2024-Q3/src/App.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/main.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/vite-env.d.ts",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/main/Card.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/main/CardsWrapper.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/searchForm/SearchForm.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/model/TypesStarWars.ts",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/pages/MainPage.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/services/ApiService.ts",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/header/Header.tsx",[],[],"/Users/mariatrejer/RS School/React/React-2024-Q3/src/components/errorBoundary/ErrorBoundary.test.tsx",[],[]]
