+++ src/components/error-boundary/error-boundary.test.tsx
@@ -1,8 +1,9 @@
 import { fireEvent, render, screen } from '@testing-library/react';
 import { expect, test } from 'vitest';
 
-import ErrorBoundary from '../components/errorBoundary/ErrorBoundary.tsx';
-import ErrorButton from '../components/header/ErrorButton.tsx';
+import ErrorButton from '../header/error-button.tsx';
+
+import ErrorBoundary from './error-boundary.tsx';
 
 test('should show fallback UI when error', () => {
   const Child = () => {
