+++ src/model/types-star-wars.ts
